Restructured project to MVVM -Moved the class files to proper folders to follow MVVM
