+++ Computer-Engineering-Capstone/MovieRatingService/src/main/java/com/example/movieratingservice/model/MovieGenre.java
@@ -1,4 +1,4 @@
-package com.example.movieratingservice;
+package com.example.movieratingservice.model;
 
 public class MovieGenre {
