+++ Computer-Engineering-Capstone/MovieRatingService/src/main/java/com/example/movieratingservice/model/Movies.java
@@ -1,4 +1,4 @@
-package com.example.movieratingservice;
+package com.example.movieratingservice.model;
 
 import java.util.ArrayList;
 import java.util.List;
