+++ Computer-Engineering-Capstone/MovieRatingService/src/main/java/com/example/movieratingservice/util/DataCallback.java
@@ -1,4 +1,4 @@
-package com.example.movieratingservice;
+package com.example.movieratingservice.util;
 
 public interface DataCallback {
     void onSuccess(String data);
